Initial Game Upload(Not original)
This is the initial game upload which is solitaire. I used both the tutorial and the unity package completed by megalmobile, link to tutorial here: https://www.megalomobile.com/lets-make-solitaire-in-unity-part-1-set-up-and-shuffle/

Disclaimer: I dont own the rights to this content and no monetary profit will come from using this content, it is only being used for educational purposes.